Moved schemas to separate file, fixed form player list error, added version and searchableProperties field in form
