+++ app/(interactive)/upload-match/page.js
@@ -1,225 +1,49 @@
-'use client';
+"use client";
 
-import React, { useState, useEffect } from 'react';
-import Form from '@rjsf/core';
-import validator from '@rjsf/validator-ajv8';
-import { useMatchData } from '../../components/MatchDataProvider.js';
-import { useAuth } from '../../components/AuthWrapper.js';
-import getTeams from '@/app/services/getTeams.js';
-import styles from '../../styles/Upload.module.css';
+import React, { useState, useEffect } from "react";
+import Form from "@rjsf/core";
+import validator from "@rjsf/validator-ajv8";
+import { useMatchData } from "../../components/MatchDataProvider.js";
+import { useAuth } from "../../components/AuthWrapper.js";
+import getTeams from "@/app/services/getTeams.js";
+import styles from "../../styles/Upload.module.css";
+import { initialSchema, uiSchema } from "@/app/services/uploadMatchSchemas.js";
 
-const parseMatchScore = (matchScore) => {
-  const sets = matchScore.split(" ");
-  const transformedSets = [];
-  
-  sets.forEach(set => {
-    const games = set.split("-");
-    const clientGamesWon = parseInt(games[0]);
-    const opponentGamesWon = parseInt(games[1]);
-    let clientTiebreakPointsWon = null;
-    let opponentTiebreakPointsWon = null;
-      
-    if (set.includes("(")) {
-      const tiebreak = set.match(/\(([^)]+)\)/)[1].split("-");
-      clientTiebreakPointsWon = parseInt(tiebreak[0]);
-      opponentTiebreakPointsWon = parseInt(tiebreak[1]);
-    }
-      
-    transformedSets.push({
-      set_number: transformedSets.length + 1,
-      clientGamesWon: clientGamesWon,
-      opponentGamesWon: opponentGamesWon,
-      clientTiebreakPointsWon: clientTiebreakPointsWon,
-      opponentTiebreakPointsWon: opponentTiebreakPointsWon
-    });
-  });
-  
-  return transformedSets;
-} 
+// const parseMatchScore = (matchScore) => {
+//   const sets = matchScore.split(" ");
+//   const transformedSets = [];
 
-const initialSchema = {
-  title: "Upload Match",
-  type: "object",
-  properties: {
-    collection: {
-      type: "string",
-      title: "Collection",
-      enum: []
-    },
-    clientTeam: {
-      type: "string",
-      title: "Client Team",
-      enum: []
-    },
-    clientPlayer: {
-      type: "string",
-      title: "Client Player",
-      enum: []
-    },
-    clientUTR: {
-      type: "string",
-      title: "Client UTR"
-    },
-    opponentTeam: {
-      type: "string",
-      title: "Opponent Team",
-      enum: []
-    },
-    opponentPlayer: {
-      type: "string",
-      title: "Opponent Player",
-      enum: []
-    },
-    opponentUTR: {
-      type: "string",
-      title: "Opponent UTR"
-    },
-    date: {
-      type: "string",
-      title: "Date",
-      format: "date"
-    },
-    matchScore: {
-      type: "object",
-      title: "Match Score",
-      properties: {
-        set1: {
-          type: "object",
-          title: "Set 1",
-          properties: {
-            clientGames: { type: "number", title: "Client Games" },
-            opponentGames: { type: "number", title: "Opponent Games" },
-            clientTiebreak: { type: "number", title: "Client Tiebreak (if applicable)" },
-            opponentTiebreak: { type: "number", title: "Opponent Tiebreak (if applicable)" }
-          }
-        },
-        set2: {
-          type: "object",
-          title: "Set 2",
-          properties: {
-            clientGames: { type: "number", title: "Client Games" },
-            opponentGames: { type: "number", title: "Opponent Games" },
-            clientTiebreak: { type: "number", title: "Client Tiebreak (if applicable)" },
-            opponentTiebreak: { type: "number", title: "Opponent Tiebreak (if applicable)" }
-          }
-        },
-        set3: {
-          type: "object",
-          title: "Set 3 (if applicable)",
-          properties: {
-            clientGames: { type: "number", title: "Client Games" },
-            opponentGames: { type: "number", title: "Opponent Games" },
-            clientTiebreak: { type: "number", title: "Client Tiebreak (if applicable)" },
-            opponentTiebreak: { type: "number", title: "Opponent Tiebreak (if applicable)" }
-          }
-        }
-      }
-    },
-    division: {
-      type: "string",
-      title: "Division",
-      enum: ["D1"]
-    },
-    event: {
-      type: "string",
-      title: "Event"
-    },
-    lineup: {
-      type: "string",
-      title: "Lineup"
-    },
-    matchVenue: {
-      type: "string",
-      title: "Match Venue"
-    },
-    round: {
-      type: "string",
-      title: "Round"
-    },
-    videoID: {
-      type: "string",
-      title: "Video ID"
-    },
-    temperature: {
-      type: "string",
-      title: "Temperature"
-    },
-    weather: {
-      type: "string",
-      title: "Weather",
-      enum: ["Cloudy", "Windy"]
-    },
-    court: {
-      type: "string",
-      title: "Court",
-      enum: ["Outdoor", "Indoor"]
-    },
-    surface: {
-      type: "string",
-      title: "Surface",
-      enum: ["Hard", "Clay", "Grass"]
-    },
-    singlesDoubles: {
-      type: "string",
-      title: "Singles/Doubles",
-      enum: ["Singles", "Doubles"]
-    },
-    jsonFile: {
-      type: "string",
-      title: "JSON File",
-      format: "data-url"
-    },
-    pdfFile: {
-      type: "string",
-      title: "PDF File",
-      format: "data-url"
-    }
-  },
-  required: ["collection", "clientTeam", "clientPlayer", "opponentTeam", "matchScore", "date"]
-};
+//   sets.forEach(set => {
+//     const games = set.split("-");
+//     const clientGamesWon = parseInt(games[0]);
+//     const opponentGamesWon = parseInt(games[1]);
+//     let clientTiebreakPointsWon = null;
+//     let opponentTiebreakPointsWon = null;
 
-const uiSchema = {
-  jsonFile: {
-    "ui:widget": "file"
-  },
-  pdfFile: {
-    "ui:widget": "file"
-  },
-  matchScore: {
-    set1: {
-      clientTiebreak: {
-        "ui:widget": "updown"
-      },
-      opponentTiebreak: {
-        "ui:widget": "updown"
-      }
-    },
-    set2: {
-      clientTiebreak: {
-        "ui:widget": "updown"
-      },
-      opponentTiebreak: {
-        "ui:widget": "updown"
-      }
-    },
-    set3: {
-      clientTiebreak: {
-        "ui:widget": "updown"
-      },
-      opponentTiebreak: {
-        "ui:widget": "updown"
-      }
-    }
-  }
-};
+//     if (set.includes("(")) {
+//       const tiebreak = set.match(/\(([^)]+)\)/)[1].split("-");
+//       clientTiebreakPointsWon = parseInt(tiebreak[0]);
+//       opponentTiebreakPointsWon = parseInt(tiebreak[1]);
+//     }
 
-export { initialSchema, uiSchema };
+//     transformedSets.push({
+//       set_number: transformedSets.length + 1,
+//       clientGamesWon: clientGamesWon,
+//       opponentGamesWon: opponentGamesWon,
+//       clientTiebreakPointsWon: clientTiebreakPointsWon,
+//       opponentTiebreakPointsWon: opponentTiebreakPointsWon
+//     });
+//   });
+
+//   return transformedSets;
+// }
 
 export default function UploadMatchForm() {
   const { createMatch } = useMatchData(); // Use the createMatch hook
   const [schema, setSchema] = useState(initialSchema);
   const [teams, setTeams] = useState([]);
   const [collections, setCollections] = useState([]);
+  const [formData, setFormData] = useState({});
 
   const { userProfile } = useAuth();
 
@@ -228,7 +52,7 @@ export default function UploadMatchForm() {
       try {
         const allTeams = await getTeams();
         setTeams(allTeams);
-        const teamNames = allTeams.map(team => team.name);
+        const teamNames = allTeams.map((team) => team.name);
 
         // Assuming userProfile.collections contains the collection names
         const userCollections = userProfile?.collections || [];
@@ -236,105 +60,140 @@ export default function UploadMatchForm() {
         setCollections(userCollections);
 
         // Update schema with team and collection names
-        setSchema(prevSchema => ({
+        setSchema((prevSchema) => ({
           ...prevSchema,
           properties: {
             ...prevSchema.properties,
             clientTeam: {
               ...prevSchema.properties.clientTeam,
-              enum: teamNames
+              enum: teamNames,
             },
             opponentTeam: {
               ...prevSchema.properties.opponentTeam,
-              enum: teamNames
+              enum: teamNames,
             },
             collection: {
               ...prevSchema.properties.collection,
-              enum: userCollections
-            }
-          }
+              enum: userCollections,
+            },
+          },
         }));
       } catch (error) {
-        console.error('Error fetching data:', error);
+        console.error("Error fetching data:", error);
       }
     };
 
-    fetchCollectionsAndTeams();    
+    fetchCollectionsAndTeams();
   }, []);
 
   const handleChange = ({ formData }) => {
+    setFormData(formData); // Prevents form reset after selecting team
     let clientPlayers = [];
     let opponentPlayers = [];
 
     if (formData.clientTeam) {
-      const selectedClientTeam = teams.find(team => team.name === formData.clientTeam);
+      const selectedClientTeam = teams.find(
+        (team) => team.name === formData.clientTeam
+      );
       if (selectedClientTeam) {
-        clientPlayers = selectedClientTeam.players?.map(player => `${player.firstName} ${player.lastName}`);
+        clientPlayers = selectedClientTeam.players?.map(
+          (player) => `${player.firstName} ${player.lastName}`
+        );
       }
     }
 
     if (formData.opponentTeam) {
-      const selectedOpponentTeam = teams.find(team => team.name === formData.opponentTeam);
+      const selectedOpponentTeam = teams.find(
+        (team) => team.name === formData.opponentTeam
+      );
       if (selectedOpponentTeam) {
-        opponentPlayers = selectedOpponentTeam.players?.map(player => `${player.firstName} ${player.lastName}`);
+        opponentPlayers = selectedOpponentTeam.players?.map(
+          (player) => `${player.firstName} ${player.lastName}`
+        );
       }
     }
 
-    setSchema(prevSchema => ({
+    setSchema((prevSchema) => ({
       ...prevSchema,
       properties: {
         ...prevSchema.properties,
         clientPlayer: {
           ...prevSchema.properties.clientPlayer,
-          enum: clientPlayers
+          enum: clientPlayers,
         },
         opponentPlayer: {
           ...prevSchema.properties.opponentPlayer,
-          enum: opponentPlayers
-        }
-      }
+          enum: opponentPlayers,
+        },
+      },
     }));
   };
 
   const handleSubmit = async ({ formData }) => {
     try {
-      const pointsJson = formData.jsonFile ? JSON.parse(atob(formData.jsonFile.split(',')[1])) : [];
+      const pointsJson = formData.jsonFile
+        ? JSON.parse(atob(formData.jsonFile.split(",")[1]))
+        : [];
       if (pointsJson.length === 0) {
-        const result = confirm("You're currently uploading an UNTAGGED match. Proceed?");
+        const result = confirm(
+          "You're currently uploading an UNTAGGED match. Proceed?"
+        );
         if (!result) throw new Error("Upload cancelled by user.");
       }
       const teams = {
         clientTeam: formData.clientTeam,
-        opponentTeam: formData.opponentTeam
+        opponentTeam: formData.opponentTeam,
       };
       const players = {
         client: {
-          firstName: formData.clientPlayer.split(' ')[0],
-          lastName: formData.clientPlayer.split(' ')[1],
-          UTR: formData.clientUTR
+          firstName: formData.clientPlayer.split(" ")[0],
+          lastName: formData.clientPlayer.split(" ")[1],
+          UTR: formData.clientUTR,
         },
         opponent: {
-          firstName: formData.opponentPlayer.split(' ')[0],
-          lastName: formData.opponentPlayer.split(' ')[1],
-          UTR: formData.opponentUTR
-        }
+          firstName: formData.opponentPlayer.split(" ")[0],
+          lastName: formData.opponentPlayer.split(" ")[1],
+          UTR: formData.opponentUTR,
+        },
       };
       const weather = {
         temperature: formData.temperature,
-        cloudy: formData.weather.includes('Cloudy'),
-        windy: formData.weather.includes('Windy')
-      }
+        cloudy: formData.weather.includes("Cloudy"),
+        windy: formData.weather.includes("Windy"),
+      };
       const matchDetails = {
-        weather, 
+        weather,
         division: formData.division,
         event: formData.event,
         lineup: formData.lineup,
         matchVenue: formData.matchVenue,
         round: formData.round,
-        indoor: formData.court === 'Indoor',
-        surface: formData.surface
-      }
-      const sets = parseMatchScore(formData.matchScore);
+        indoor: formData.court === "Indoor",
+        surface: formData.surface,
+      };
+      // const sets = parseMatchScore(formData.matchScore);
+      const sets = [
+        formData.matchScore.set1,
+        formData.matchScore.set2,
+        ...(formData.matchScore.set3 ? formData.matchScore.set3 : []),
+      ];
+
+      //Searchable properties
+      const searchableProperties = [
+        "clientTeam",
+        "clientPlayer",
+        "opponentTeam",
+        "opponentPlayer",
+        "date",
+        "division",
+        "event",
+        "lineup",
+        "matchVenue",
+        "round",
+        "court",
+        "surface",
+        "singlesDoubles",
+      ];
 
       // Use the createMatch hook to upload the match
       await createMatch(formData.collection, {
@@ -345,11 +204,13 @@ export default function UploadMatchForm() {
         teams,
         players,
         matchDate: formData.date,
-        singles: formData.singlesDoubles === 'Singles',
-        matchDetails
+        singles: formData.singlesDoubles === "Singles",
+        matchDetails,
+        searchableProperties,
+        version: "v1", // Current version for new matches added
       });
 
-      alert('Match uploaded successfully!');
+      alert("Match uploaded successfully!");
     } catch (error) {
       console.error("Error uploading match:", error);
       alert(`Error uploading match: ${error.message}`);
@@ -360,16 +221,19 @@ export default function UploadMatchForm() {
     <div className={styles.container}>
       <div>
         <h1 className={styles.title}>Upload Match</h1>
-        <h3>Make sure you add the player in &apos;Upload Team&apos; before this!</h3>
+        <h3>
+          Make sure you add the player in &apos;Upload Team&apos; before this!
+        </h3>
         <Form
           key={JSON.stringify(schema)}
           schema={schema}
           uiSchema={uiSchema}
+          formData={formData}
           onChange={handleChange}
           onSubmit={handleSubmit}
           validator={validator}
         />
       </div>
     </div>
-  );  
+  );
 }
